refatoração: separando código em módulos (menus, eleitor, db, validacoes)
